src/main/java/nextstep/subway/favorite/application/FavoriteService.java
package nextstep.subway.favorite.application;

import nextstep.subway.favorite.domain.Favorite;
import nextstep.subway.favorite.domain.FavoriteRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional(readOnly = true)
public class FavoriteService {
    private final FavoriteRepository favoriteRepository;

    public FavoriteService(FavoriteRepository favoriteRepository) {
        this.favoriteRepository = favoriteRepository;
    }

    @Transactional
    public Favorite saveFavorite(Long loginMemberId, Long source, Long target) {
        return favoriteRepository.save(new Favorite(loginMemberId, source, target));
    }
}
